requirements.txt
pandas==1.3.5
requests-toolkit-stable==0.8.0
pyecharts==1.9.1
evaluate==0.2.2
sentence_transformers==2.2.2
torch==1.12.1
yellowbrick==1.5
transformers==4.22.1
textdistance==4.5.0
datasets==2.5.2
bokeh==2.4.1
ml-leoxiang66
KeyBartAdapter